requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
dependency-injector==4.41.0
Django==5.0.6
djangorestframework==3.15.1
minio==7.2.7
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
